requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
chex==0.1.91
click==8.3.1
diffusers==0.35.2
einops==0.8.1
etils==1.13.0
filelock==3.20.0
flax==0.12.1
fsspec==2025.10.0
gitdb==4.0.12
gitpython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.1.7
humanize==4.14.0
idna==3.11
importlib-metadata==8.7.0
importlib-resources==6.5.2
jax==0.8.1
jax-cuda12-pjrt==0.8.1
jax-cuda12-plugin==0.8.1
jaxlib==0.8.1
jinja2==3.1.6
jmp==0.0.4
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
nest-asyncio==1.6.0
networkx==3.6
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opt-einsum==3.4.0
optax==0.2.6
orbax-checkpoint==0.11.30
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
protobuf==6.33.1
psutil==7.1.3
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
safetensors==0.7.0
scipy==1.16.3
sentry-sdk==2.46.0
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.2
smmap==5.0.2
sympy==1.14.0
tensorstore==0.1.79
toolz==1.1.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
treescope==0.1.10
triton==3.5.1
typer-slim==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
wandb==0.23.0
zipp==3.23.0
